docs: Improvement in Translation, Package.json and GitIgnore (#84)
Version Update in Package.json, Addition in gitignore, Updates in Translation and Implementation of Routes for Translation

Sorry for not being able to integrate the cli, I can not run my machine. Can you please add me as documentation and code? It seems that this PR is breaking some things like I can fix?